UPDATE: combine all rework migrations into one

+++ Src/Infrastructure/Ws.Database.Nhibernate/Entities/Print/Labels/SqlLabelMap.cs
@@ -101,7 +101,7 @@ internal sealed class SqlLabelMap : ClassMapping<Label>
         Property(x => x.BundleCount, m =>
         {
             m.Column("BUNDLE_COUNT");
-            m.Type(NHibernateUtil.Int32);
+            m.Type(NHibernateUtil.UInt16);
             m.NotNullable(true);
         });
